php commit

+++ delete.php
@@ -1,10 +1,11 @@
 <?php
-include("conn.php")
-$id = $_GET['id'];
-$delete ="DELETE FROM patient WHERE id = ?";
-$stmt = $conn -> prepare($delete);
-$stmt -> execute([$id]);
-header("Location:select.php");
+include("conn.php");
+if (isset($_GET['id'])) {
+    $id = $_GET['id'];
+    $delete ="DELETE FROM patient WHERE id = ?";
+    $stmt = $conn -> prepare($delete);
+    $stmt -> execute([$id]);
+    header("Location:select.php");
 
-
+}
 ?>
